requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
peewee==3.16.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
six==1.16.0
Werkzeug==2.3.4
zipp==3.15.0
